Pull request review     python: conversion. rosbag add progress bar     python: convert script update documentation and argument since now it is not ros topic but any topic name (ros or rock)     python: remove unnecessary file.     python: base accumulator class in data module for two types of EventAccumulators (ros and rock/pocolog)     README: complete documentation with Dockerfile
