+++ README.md
@@ -28,8 +28,9 @@ Second, installation of packages in a conda environment to run the reconstructio
 ### Conversion to H5
 Our current conversion code supports 3 event file formats:
 1. Rosbags with [dvs\_msgs](https://github.com/uzh-rpg/rpg_dvs_ros/tree/master/dvs_msgs)
-2. Prophesee raw format using [Metavision 2.2](https://docs.prophesee.ai/2.2.0/installation/index.html)
-2. Prophesee dat format using [Metavision 2.X](https://docs.prophesee.ai/stable/data_formats/file_formats/dat.html)
+2. Pocolog with [base/samples/EventArray](https://github.com/rock-core/base-orogen-types)
+3. Prophesee raw format using [Metavision 2.2](https://docs.prophesee.ai/2.2.0/installation/index.html)
+4. Prophesee dat format using [Metavision 2.X](https://docs.prophesee.ai/stable/data_formats/file_formats/dat.html)
 
 Regardeless of the event file format:
 ```bash
@@ -87,6 +88,37 @@ If you have an older Metavision version (for example Metavision 2.0), first conv
 ### Conversion to H5 from ROCK logs
 
 The [conversion script](https://github.com/uzh-rpg/e2calib/blob/main/python/pocolog_convert.py) simply requires the path to the pocolog/rock file and the name of the outport (topic) of the events. In case of not output_file provided it will generate a .h5 file in the same folder where the pocolog file is located.
+
+### Conversion to H5 from Pocolog files
+
+Pocolog is the file format for logging data in [ROCK](https://www.rock-robotics.org/). It is equivalent to the bag format for ROS. More specifically [Pocolog](https://github.com/rock-core/tools-pocolog) is the tool to manupilate [log](https://github.com/rock-core/tools-logger) files.
+
+The [conversion script](https://github.com/uzh-rpg/e2calib/blob/main/python/convert.py) understands the Pocolog file format from the pocolog [conversion](https://github.com/uzh-rpg/e2calib/blob/main/python/conversion/pocolog.py) file. You need a ROCK installation with the Pocolog python bindings [Pocolog Pybind](https://github.com/jhidalgocarrio/tools-pocolog_pybind) installed in order to convert events in pocolog to h5 format. Please follow the installation guide to install ROCK on your system: [How to install ROCK](https://www.rock-robotics.org/documentation/installation.html). Afterwards clone the Pocolog Python bindings:
+
+```bash
+git clone git@github.com:jhidalgocarrio/tools-pocolog_pybind.git <rock-path>/tools/pocolog_pybind
+```
+
+Compile and install the Pocolog Python bindings:
+
+```bash
+source <rock-path>/env.sh
+cd <rock-path>/tools/pocolog_pybind
+amake
+python3 -m pip install <rock-path>/tools/pocolog_pybind
+```
+
+Afterward, you can simple use the [conversion script](https://github.com/uzh-rpg/e2calib/blob/main/python/convert.py) with the path to the pocolog file and the port name  (i.e.: similar to ros topic name) containing the events (e.g.: --topic /camera_prophesee.events).
+
+```bash
+python convert.py <pocolog_file> -t <port_name>
+```
+
+Alternativelly, you can also use our [Dockerfile](https://download.ifi.uzh.ch/rpg/e2calib/pocolog/Dockerfile) to create a docker image which has all the tools installed. You will be able to convert from pocolog to rosbag and h5 format. To build the docker image run:
+
+```bash
+docker build -t <image_name> -f Dockerfile .
+```
 
 ### Reconstruction
