+++ python/conversion/ros.py
@@ -1,32 +1,10 @@
 from pathlib import Path
 
 import numpy as np
+import tqdm
 import rosbag
 
-from data.format import Events
-
-class EventAccumulator:
-    def __init__(self):
-        self.x = list()
-        self.y = list()
-        self.p = list()
-        self.t = list()
-
-    def add_event(self, event):
-        self.x.append(event.x)
-        self.y.append(event.y)
-        self.p.append(int(event.polarity))
-        # floor to microseconds.
-        self.t.append(event.ts.to_nsec()//1000)
-
-    def get_events(self):
-        events = Events(
-                np.asarray(self.x, dtype='uint16'),
-                np.asarray(self.y, dtype='uint16'),
-                np.asarray(self.p, dtype='uint8'),
-                np.asarray(self.t, dtype='int64'))
-        return events
-
+from data.accumulator import EventAccumulatorRos
 
 def ev_generator(bagpath: Path, delta_t_ms: int=1000, topic: str='/dvs/events'):
     assert bagpath.exists()
@@ -35,11 +13,12 @@ def ev_generator(bagpath: Path, delta_t_ms: int=1000, topic: str='/dvs/events'):
     delta_t_ns = delta_t_ms * 10**6
 
     t_ev_acc_end_ns = None
-    ev_acc = EventAccumulator()
+    ev_acc = EventAccumulatorRos()
 
     init = False
     last_time = 0
     with rosbag.Bag(str(bagpath), 'r') as bag:
+        pbar = tqdm.tqdm(total=bag.get_message_count(topic))
         for topic, msg, ros_time in bag.read_messages(topics=[topic]):
             if not init:
                 init = True
@@ -55,5 +34,6 @@ def ev_generator(bagpath: Path, delta_t_ms: int=1000, topic: str='/dvs/events'):
                     events = ev_acc.get_events()
                     yield events
                     t_ev_acc_end_ns = t_ev_acc_end_ns + delta_t_ns
-                    ev_acc = EventAccumulator()
+                    ev_acc = EventAccumulatorRos()
                     ev_acc.add_event(event)
+            pbar.update(1)
