+++ python/convert.py
@@ -9,7 +9,7 @@ if __name__ == '__main__':
     parser = argparse.ArgumentParser('Convert events to h5 format to prepare for calibration.')
     parser.add_argument('input_file', help='Path to file which will be converted to hdf5 format.')
     parser.add_argument('--output_file', '-o', default="", help='Output path for h5 file. Default: Input path but with h5 suffix.')
-    parser.add_argument('--ros_topic', '-rt', default='/dvs/events', help='ROS topic for events if input file is a rosbag.')
+    parser.add_argument('--topic', '-t', default='/dvs/events', help='Topic name for events if input file is a rosbag(ROS) or pocolog(ROCK).')
 
     args = parser.parse_args()
 
@@ -22,9 +22,9 @@ if __name__ == '__main__':
         output_file = Path(input_file).parent / (input_file.stem + '.h5')
     assert not output_file.exists(), f"{output_file} already exists."
 
-    rostopic = args.ros_topic
+    topic = args.topic
 
-    event_generator = conversion.format.get_generator(input_file, delta_t_ms=1000, topic=rostopic)
+    event_generator = conversion.format.get_generator(input_file, delta_t_ms=1000, topic=topic)
     h5writer = conversion.h5writer.H5Writer(output_file)
 
     for events in event_generator():
